feat: introduce new API crate with initial implementation, including routing, middleware, and health check endpoints, along with configuration for logging and authentication

+++ crates/config/src/queue.rs
@@ -4,7 +4,7 @@ use serde::Deserialize;
 #[derive(Debug, Clone, Deserialize)]
 pub struct QueueConfig {
     /// Number of worker threads to process jobs
-    pub worker_count: u16,
+    pub worker_count: usize,
     /// Maximum number of jobs in the queue
     pub max_queue_size: usize,
 }
